nvim: Add a way to track if we're reloading config
Certain settings, like tabstop, we want to set when we first load to
overwrite the default. However, if we set them when reloading the
config, they may overwrite buffer-local values that were set via after/.

Here, we provide is_first_load() which we can use to ensure these are
only set on startup.